refactors user info into it's own component
